Moved tests around, removed a test that is now a duplicate
